Checkpoint

+++ src/com/cjwatts/auctionsystem/gui/BidEvent.java
@@ -5,17 +5,17 @@ import java.awt.AWTEvent;
 public class BidEvent extends AWTEvent {
 	private static final long serialVersionUID = 1L;
 
-	private float bid;
+	private double bid;
 	
 	public BidEvent(Object source, int id) {
 		super(source, id);
 	}
 
-	public float getBid() {
+	public double getBid() {
 		return bid;
 	}
 
-	public void setBid(float bid) {
+	public void setBid(double bid) {
 		this.bid = bid;
 	}
 }
